WIP: Issue-#126      fix broken tests (wip)      rename JCUnit8X -> JCUnit8
