+++ src/main/java/com/github/dakusui/jcunit8/runners/junit4/JCUnit8.java
@@ -1,42 +1,41 @@
 package com.github.dakusui.jcunit8.runners.junit4;
 
 import com.github.dakusui.jcunit.core.tuples.Tuple;
+import com.github.dakusui.jcunit.core.utils.Checks;
 import com.github.dakusui.jcunit8.core.Utils;
 import com.github.dakusui.jcunit8.exceptions.TestDefinitionException;
 import com.github.dakusui.jcunit8.factorspace.Constraint;
-import com.github.dakusui.jcunit8.factorspace.TestPredicate;
+import com.github.dakusui.jcunit8.factorspace.ParameterSpace;
+import com.github.dakusui.jcunit8.pipeline.Config;
+import com.github.dakusui.jcunit8.pipeline.Pipeline;
 import com.github.dakusui.jcunit8.pipeline.stages.ConfigFactory;
 import com.github.dakusui.jcunit8.runners.core.NodeUtils;
-import com.github.dakusui.jcunit8.runners.junit4.annotations.AfterTestCase;
-import com.github.dakusui.jcunit8.runners.junit4.annotations.BeforeTestCase;
-import com.github.dakusui.jcunit8.runners.junit4.annotations.ConfigureWith;
+import com.github.dakusui.jcunit8.runners.junit4.annotations.*;
 import com.github.dakusui.jcunit8.runners.junit4.utils.InternalUtils;
-import com.github.dakusui.jcunit8.testsuite.TestCase;
-import com.github.dakusui.jcunit8.testsuite.TestSuite;
-import org.junit.*;
-import org.junit.internal.runners.rules.RuleMemberValidator;
-import org.junit.internal.runners.statements.InvokeMethod;
-import org.junit.internal.runners.statements.RunAfters;
-import org.junit.internal.runners.statements.RunBefores;
+import com.github.dakusui.jcunit8.testsuite.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
 import org.junit.runner.Description;
 import org.junit.runner.Runner;
 import org.junit.runner.notification.RunNotifier;
-import org.junit.runners.BlockJUnit4ClassRunner;
 import org.junit.runners.ParentRunner;
 import org.junit.runners.model.FrameworkMethod;
 import org.junit.runners.model.InitializationError;
 import org.junit.runners.model.Statement;
 import org.junit.runners.model.TestClass;
+import org.junit.validator.AnnotationsValidator;
+import org.junit.validator.PublicClassValidator;
 import org.junit.validator.TestClassValidator;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 import static com.github.dakusui.jcunit8.core.Utils.createTestClassMock;
 import static com.github.dakusui.jcunit8.exceptions.FrameworkException.unexpectedByDesign;
@@ -44,63 +43,107 @@ import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
 
 public class JCUnit8 extends org.junit.runners.Parameterized {
-
-  private final TestSuite    testSuite;
   private final List<Runner> runners;
+  private final TestSuite    testSuite;
 
+  /**
+   * Only called reflectively. Do not use programmatically.
+   *
+   * @param klass A test class
+   */
   public JCUnit8(Class<?> klass) throws Throwable {
     super(klass);
-    ConfigFactory configFactory = getConfigFactory();
-    TestClass parameterSpaceDefinition = createParameterSpaceDefinitionTestClass();
-    Collection<String> involvedParameterNames = InternalUtils.involvedParameters(new TestClass(klass));
-    this.testSuite = JCUnit8X.buildTestSuite(
-        configFactory.create(),
-        JCUnit8X.buildParameterSpace(
-            new ArrayList<>(
-                JCUnit8X.buildParameterMap(parameterSpaceDefinition).values()
-            ).stream(
-            ).filter(
-                parameter -> involvedParameterNames.contains(parameter.getName())
-            ).collect(
-                toList()
-            ),
-            NodeUtils.allTestPredicates(getTestClass()).values().stream()
-                .filter(each -> each instanceof Constraint)
-                .map(Constraint.class::cast)
-                .collect(toList())
-        ),
-        TestScenarioFactoryForJUnit4.create(new TestClass(klass))
-    );
-    this.runners = createRunners();
+    this.runners = createRunners(this.testSuite = buildTestSuite(
+        getTestClass(),
+        createParameterSpaceDefinitionTestClass(),
+        getConfigFactory()
+    ));
   }
 
-  public static Statement createMethodInvoker(FrameworkMethod method, TestCaseRunner testCaseRunner, Object test) throws Throwable {
-    return new InvokeMethod(method, test) {
-      @Override
-      public void evaluate() throws Throwable {
-        InternalUtils.invokeExplosivelyWithArgumentsFromTestInput(method, testCaseRunner.getTestCase().getTestInput());
+  private static TestClassValidator[] createValidatorsFor(TestClass parameterSpaceDefinitionClass) {
+    return new TestClassValidator[] {
+        new AnnotationsValidator(),
+        new PublicClassValidator(),
+        new TestClassValidator() {
+          @Override
+          public List<Exception> validateTestClass(TestClass testClass) {
+            return new LinkedList<Exception>() {
+              {
+                validateFromAnnotationsAreReferencingExistingParameterSourceMethods(BeforeTestCase.class, testClass, this);
+                validateFromAnnotationsAreReferencingExistingParameterSourceMethods(Before.class, testClass, this);
+                validateFromAnnotationsAreReferencingExistingParameterSourceMethods(Test.class, testClass, this);
+                validateFromAnnotationsAreReferencingExistingParameterSourceMethods(After.class, testClass, this);
+                validateFromAnnotationsAreReferencingExistingParameterSourceMethods(AfterTestCase.class, testClass, this);
+                validateAtLeastOneTestMethod(testClass, this);
+              }
+
+            };
+          }
+
+          private void validateAtLeastOneTestMethod(TestClass testClass, LinkedList<Exception> errors) {
+            if (testClass.getAnnotatedMethods(Test.class).isEmpty()) {
+              errors.add(new Exception("No runnable methods"));
+            }
+          }
+
+          private void validateFromAnnotationsAreReferencingExistingParameterSourceMethods(Class<? extends Annotation> ann, TestClass testClass, List<Exception> errors) {
+            testClass.getAnnotatedMethods(ann)
+                .forEach(
+                    frameworkMethod -> Stream.of(frameworkMethod.getMethod().getParameterAnnotations())
+                        .forEach((Annotation[] annotations) -> Stream.of(annotations)
+                            .filter((Annotation annotation) -> annotation instanceof From)
+                            .forEach((Annotation annotation) -> {
+                              List<FrameworkMethod> methods = parameterSpaceDefinitionClass.getAnnotatedMethods(ParameterSource.class).stream()
+                                  .filter(
+                                      (FrameworkMethod each) ->
+                                          Objects.equals(each.getName(), From.class.cast(annotation).value()))
+                                  .collect(toList());
+                              if (methods.isEmpty())
+                                errors.add(new Exception(
+                                    format(
+                                        "A method '%s' annotated with '%s' is not defined in '%s'",
+                                        From.class.cast(annotation).value(),
+                                        ParameterSource.class.getSimpleName(),
+                                        parameterSpaceDefinitionClass.getJavaClass().getCanonicalName()
+                                    )));
+                            })));
+          }
+        }
+    };
+  }
+
+  private static ParameterSpace buildParameterSpace(List<com.github.dakusui.jcunit8.factorspace.Parameter> parameters, List<Constraint> constraints) {
+    return new ParameterSpace.Builder()
+        .addAllParameters(parameters)
+        .addAllConstraints(constraints)
+        .build();
+  }
+
+  private static TestSuite buildTestSuite(Config config, ParameterSpace parameterSpace, TestScenario testScenario) {
+    return Pipeline.Standard.<Tuple>create().execute(config, parameterSpace, testScenario);
+  }
+
+  private static SortedMap<String, com.github.dakusui.jcunit8.factorspace.Parameter> buildParameterMap(TestClass parameterSpaceDefinitionTestClass) {
+    return new TreeMap<String, com.github.dakusui.jcunit8.factorspace.Parameter>() {
+      {
+        parameterSpaceDefinitionTestClass.getAnnotatedMethods(ParameterSource.class).forEach(
+            frameworkMethod -> put(frameworkMethod.getName(),
+                buildParameterFactoryCreatorFrom(frameworkMethod)
+                    .apply(Utils.createInstanceOf(parameterSpaceDefinitionTestClass))
+                    .create(frameworkMethod.getName())
+            ));
       }
     };
   }
 
-
-  @Override
-  protected void collectInitializationErrors(List<Throwable> errors) {
-    this.applyValidators(errors);
-  }
-
-  private void applyValidators(List<Throwable> errors) {
-    if (getTestClass().getJavaClass() != null) {
-      for (TestClassValidator each : JCUnit8X.createValidatorsFor(createParameterSpaceDefinitionTestClass())) {
-        errors.addAll(each.validateTestClass(getTestClass()));
+  private static Function<Object, com.github.dakusui.jcunit8.factorspace.Parameter.Factory> buildParameterFactoryCreatorFrom(FrameworkMethod method) {
+    return (Object o) -> {
+      try {
+        return (com.github.dakusui.jcunit8.factorspace.Parameter.Factory) method.invokeExplosively(o);
+      } catch (Throwable throwable) {
+        throw unexpectedByDesign(throwable);
       }
-    }
-  }
-
-
-  @Override
-  protected List<Runner> getChildren() {
-    return this.runners;
+    };
   }
 
   /**
@@ -109,6 +152,88 @@ public class JCUnit8 extends org.junit.runners.Parameterized {
   @Override
   protected TestClass createTestClass(Class<?> testClass) {
     return createTestClassMock(super.createTestClass(testClass));
+  }
+
+  @Override
+  protected List<Runner> getChildren() {
+    return this.runners;
+  }
+
+  @Override
+  protected void collectInitializationErrors(List<Throwable> errors) {
+    this.applyValidators(errors);
+  }
+
+  protected Statement withBeforeClasses(Statement statement) {
+    return this.testSuite.getScenario().preSuiteProcedures().isEmpty() ?
+        statement :
+        InternalUtils.createRunBeforesForTestInput(
+            statement,
+            this.testSuite.getScenario().preSuiteProcedures(),
+            Tuple.builder().put("@suite", this.testSuite).build()
+        );
+  }
+
+
+  protected Statement withAfterClasses(Statement statement) {
+    return this.testSuite.getScenario().preSuiteProcedures().isEmpty() ?
+        statement :
+        InternalUtils.createRunAftersForTestInput(
+            statement,
+            this.testSuite.getScenario().postSuiteProcedures(),
+            Tuple.builder().put("@suite", this.testSuite).build()
+        );
+  }
+
+
+  public static TestSuite buildTestSuite(
+      TestClass testClass,
+      TestClass parameterSpaceDefinitionTestClass,
+      ConfigFactory configFactory
+  ) {
+    Collection<String> involvedParameterNames = InternalUtils.involvedParameters(testClass);
+    return buildTestSuite(
+        configFactory.create(),
+        buildParameterSpace(
+            new ArrayList<>(
+                buildParameterMap(parameterSpaceDefinitionTestClass).values()
+            ).stream(
+            ).filter(
+                parameter -> involvedParameterNames.contains(parameter.getName())
+            ).collect(
+                toList()
+            ),
+            NodeUtils.allTestPredicates(testClass).values().stream()
+                .filter(each -> each instanceof Constraint)
+                .map(Constraint.class::cast)
+                .collect(toList())
+        ),
+        TestScenarioFactoryForJUnit4.create(testClass)
+    );
+  }
+
+  private List<Runner> createRunners(TestSuite testSuite) {
+    return IntStream.range(
+        0, testSuite.size()
+    ).mapToObj(
+        i -> {
+          try {
+            return new TestCaseRunner(this.getTestClass().getJavaClass(), i, testSuite);
+          } catch (InitializationError initializationError) {
+            throw Checks.wrap(initializationError);
+          }
+        }
+    ).collect(
+        toList()
+    );
+  }
+
+  private void applyValidators(List<Throwable> errors) {
+    if (getTestClass().getJavaClass() != null) {
+      for (TestClassValidator each : createValidatorsFor(createParameterSpaceDefinitionTestClass())) {
+        errors.addAll(each.validateTestClass(getTestClass()));
+      }
+    }
   }
 
   private ConfigFactory getConfigFactory() {
@@ -133,45 +258,20 @@ public class JCUnit8 extends org.junit.runners.Parameterized {
     return ret;
   }
 
-  private List<Runner> createRunners() {
-    AtomicInteger i = new AtomicInteger(0);
-    return this.testSuite.stream()
-        .map((Function<TestCase, Runner>) tupleTestCase -> {
-          try {
-            return new TestCaseRunner(i.getAndIncrement(), tupleTestCase, NodeUtils.allTestPredicates(getTestClass()));
-          } catch (InitializationError initializationError) {
-            throw unexpectedByDesign(formatInitializationErrorMessage(initializationError));
-          }
-        })
-        .collect(toList());
-  }
+  private static class TestCaseRunner extends ParentRunner<TestOracle> implements ITestCaseRunner {
 
-  private static String formatInitializationErrorMessage(InitializationError e) {
-    return e.getCauses().stream().map(Throwable::getMessage).collect(Collectors.joining());
-  }
+    private final int       id;
+    private final TestSuite testSuite;
 
-  /**
-   * This method is only used through reflection to let JUnit know the test case is ignored since
-   * no matching test method is defined for it.
-   *
-   * @see TestCaseRunner#getDummyMethodForNoMatchingMethodFound()
-   */
-  @Ignore
-  @SuppressWarnings({ "unused", "WeakerAccess" })
-  public static void noMatchingTestMethodIsFoundForThisTestCase() {
-  }
-
-
-  public class TestCaseRunner extends BlockJUnit4ClassRunner implements ITestCaseRunner {
-    private final TestCase tupleTestCase;
-    int id;
-    private final SortedMap<String, TestPredicate> predicates;
-
-    TestCaseRunner(int id, TestCase tupleTestCase, SortedMap<String, TestPredicate> predicates) throws InitializationError {
-      super(JCUnit8.this.getTestClass().getJavaClass());
-      this.tupleTestCase = tupleTestCase;
+    /**
+     * Constructs a new {@code ParentRunner} that will run {@code @TestClass}
+     *
+     * @param javaClass A class that defines a test suite to be run.
+     */
+    private TestCaseRunner(Class<?> javaClass, int id, TestSuite testSuite) throws InitializationError {
+      super(javaClass);
       this.id = id;
-      this.predicates = predicates;
+      this.testSuite = testSuite;
     }
 
     @Override
@@ -180,55 +280,33 @@ public class JCUnit8 extends org.junit.runners.Parameterized {
     }
 
     @Override
-    protected String testName(final FrameworkMethod method) {
-      return format("%s[%d]", method.getName(), this.id);
+    protected List<TestOracle> getChildren() {
+      return testSuite.getScenario().oracles();
     }
 
     @Override
-    protected void validateConstructor(List<Throwable> errors) {
-      validateZeroArgConstructor(errors);
-    }
-
-    @Override
-    protected void validateTestMethods(List<Throwable> errors) {
-    }
-
-    @Override
-    protected Description describeChild(FrameworkMethod method) {
+    protected Description describeChild(TestOracle child) {
       return Description.createTestDescription(
           getTestClass().getJavaClass(),
-          testName(method),
-          method.getAnnotations()
+          String.format("%s[%s]", child.getName(), this.id),
+          new Annotation[0]
       );
     }
 
     @Override
-    public List<FrameworkMethod> getChildren() {
-      try {
-        List<FrameworkMethod> ret = new LinkedList<>();
-        for (FrameworkMethod each : computeTestMethods()) {
-          if (shouldInvoke(each, tupleTestCase.getTestInput()))
-            ret.add(each);
-        }
-        if (ret.isEmpty())
-          ret.add(getDummyMethodForNoMatchingMethodFound());
-        return ret;
-      } catch (Throwable t) {
-        throw unexpectedByDesign(t);
-      }
-    }
+    protected void runChild(TestOracle child, RunNotifier notifier) {
+      Description description = describeChild(child);
 
-    public TestCase getTestCase() {
-      return tupleTestCase;
+      Tuple testInput = composeTestInput(this.testSuite.get(this.id).getTestInput());
+      if (!child.shouldInvoke().test(testInput)) {
+        notifier.fireTestIgnored(description);
+      } else {
+        runLeaf(oracleBlock(child, testInput), description, notifier);
+      }
     }
 
     @Override
     protected void collectInitializationErrors(List<Throwable> errors) {
-      validatePublicVoidNoArgMethods(BeforeClass.class, true, errors);
-      validatePublicVoidNoArgMethods(AfterClass.class, true, errors);
-      RuleMemberValidator.CLASS_RULE_VALIDATOR.validate(getTestClass(), errors);
-      RuleMemberValidator.CLASS_RULE_METHOD_VALIDATOR.validate(getTestClass(), errors);
-      applyValidators(errors);
     }
 
     @Override
@@ -241,88 +319,16 @@ public class JCUnit8 extends org.junit.runners.Parameterized {
       return statement;
     }
 
-    @Override
-    protected Statement withBefores(FrameworkMethod method, Object target,
-        Statement statement) {
-      List<FrameworkMethod> befores = getTestClass().getAnnotatedMethods(Before.class).stream(
-      ).map(
-          InternalUtils.frameworkMethodInvokingArgumentsFromTestCase_(this, target)
-      ).collect(
-          toList()
-      );
-      return befores.isEmpty() ?
-          statement :
-          new RunBefores(statement, befores, target);
-    }
-
-    @Override
-    protected Statement withAfters(FrameworkMethod method, Object target,
-        Statement statement) {
-      List<FrameworkMethod> afters = getTestClass().getAnnotatedMethods(After.class).stream(
-      ).map(
-          InternalUtils.frameworkMethodInvokingArgumentsFromTestCase_(this, target)
-      ).collect(
-          toList()
-      );
-      return afters.isEmpty() ?
-          statement :
-          new RunAfters(statement, afters, target);
-    }
-
-    /**
-     * Returns a {@link Statement} that invokes {@code method} on {@code test}
-     */
-    @Override
-    protected Statement methodInvoker(final FrameworkMethod method, final Object test) {
-      try {
-        return createMethodInvoker(method, this, test);
-      } catch (Throwable throwable) {
-        throw new Error(throwable);
-      }
-    }
-
-    @Override
-    public Object createTest() {
-      return Utils.createInstanceOf(getTestClass());
-    }
-
-    @Override
-    protected Annotation[] getRunnerAnnotations() {
-      return new Annotation[0];
-    }
-
-    private boolean shouldInvoke(FrameworkMethod method, Tuple tuple) {
-      return shouldInvoke(method).test(tuple);
-    }
-
-    private Predicate<Tuple> shouldInvoke(FrameworkMethod method) {
-      return tuple -> InternalUtils.shouldInvoke(method, predicates).test(tuple);
-    }
-
-    private FrameworkMethod getDummyMethodForNoMatchingMethodFound() {
-      try {
-        return new FrameworkMethod(JCUnit8.class.getMethod("noMatchingTestMethodIsFoundForThisTestCase")) {
-          @Override
-          public String getName() {
-
-            return format("%s:%s", super.getName(), Objects.toString(tupleTestCase));
-          }
-        };
-      } catch (NoSuchMethodException e) {
-        throw unexpectedByDesign(e);
-      }
-    }
-
     private Statement withBeforeTestCases(Statement statement) {
-      List<FrameworkMethod> befores = JCUnit8.this.getTestClass().getAnnotatedMethods(BeforeTestCase.class);
-      return befores.isEmpty() ? statement :
-          InternalUtils.createRunBeforesForTestCase_(statement, befores, this);
+      return testSuite.getScenario().preTestInputProcedures().isEmpty() ?
+          statement :
+          InternalUtils.createRunBeforesForTestInput(statement, testSuite.getScenario().preTestInputProcedures(), this.getTestCase().getTestInput());
     }
 
     private Statement withAfterTestCases(Statement statement) {
       List<FrameworkMethod> afters = getTestClass().getAnnotatedMethods(AfterTestCase.class);
       return afters.isEmpty() ? statement :
-          InternalUtils.createRunAftersForTestCase_(statement, afters, this);
+          InternalUtils.createRunAftersForTestInput(statement, testSuite.getScenario().postTestInputProcedures(), this.getTestCase().getTestInput());
     }
 
     private boolean checkIfAllChildrenAreIgnored() {
@@ -339,6 +345,66 @@ public class JCUnit8 extends org.junit.runners.Parameterized {
         throw new Error(e);
       }
     }
-  }
 
+    private Tuple composeTestInput(Tuple tuple) {
+      try {
+        return Tuple.builder()
+            .putAll(tuple)
+            .put("@ins", getTestClass().getOnlyConstructor().newInstance())
+            .put("@suite", testSuite)
+            .build();
+      } catch (InstantiationException | InvocationTargetException | IllegalAccessException e) {
+        throw Checks.wrap(e);
+      }
+    }
+
+    private Statement oracleBlock(TestOracle testOracle, Tuple testInput) {
+      Statement statement = oracleInvoker(testOracle, testInput);
+      statement = withBeforesForTestOracle(testInput, statement);
+      statement = withAftersForTestOracle(testInput, statement);
+      return statement;
+    }
+
+    private Statement oracleInvoker(TestOracle oracle, Tuple testInput) {
+      return new Statement() {
+        @Override
+        public void evaluate() throws Throwable {
+          oracle.accept(testInput);
+        }
+      };
+    }
+
+    private Statement withBeforesForTestOracle(Tuple testInput, Statement statement) {
+      List<TupleConsumer> befores = testSuite.getScenario().preOracleProcedures();
+      return befores.isEmpty() ?
+          statement :
+          new Statement() {
+            @Override
+            public void evaluate() throws Throwable {
+              for (Consumer<Tuple> before : befores)
+                before.accept(testInput);
+              statement.evaluate();
+            }
+          };
+    }
+
+    private Statement withAftersForTestOracle(Tuple testInput, Statement statement) {
+      List<TupleConsumer> afters = testSuite.getScenario().postOracleProcedures();
+      return afters.isEmpty() ?
+          statement :
+          new Statement() {
+            @Override
+            public void evaluate() throws Throwable {
+              statement.evaluate();
+              for (Consumer<Tuple> after : afters)
+                after.accept(testInput);
+            }
+          };
+    }
+
+    @Override
+    public TestCase getTestCase() {
+      return this.testSuite.get(this.id);
+    }
+  }
 }
