+++ src/main/java/com/github/dakusui/jcunit8/testsuite/TestOracle.java
@@ -6,82 +6,10 @@ import com.github.dakusui.jcunit.core.utils.Checks;
 import java.io.IOException;
 import java.util.Formattable;
 import java.util.Formatter;
-import java.util.Objects;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
 public interface TestOracle extends TupleConsumer {
-  class Builder {
-    private Predicate<Tuple> shouldInvoke = new Predicate<Tuple>() {
-      @Override
-      public boolean test(Tuple tuple) {
-        return true;
-      }
-
-      @Override
-      public String toString() {
-        return "always";
-      }
-    };
-    private Function<Tuple, Result> when;
-    private Predicate<Result> then      = new Predicate<Result>() {
-      @Override
-      public boolean test(Result result) {
-        return result.exitedWith().equals(Result.Exit.RETURNING_VALUE);
-      }
-
-      @Override
-      public String toString() {
-        return "valueReturned";
-      }
-    };
-    private Assertion         assertion = new Assertion() {
-    };
-
-    public Builder() {
-    }
-
-    public Builder shouldInvoke(Predicate<Tuple> shouldInvoke) {
-      this.shouldInvoke = Objects.requireNonNull(shouldInvoke);
-      return this;
-    }
-
-    public Builder when(Function<Tuple, Result> when) {
-      this.when = Objects.requireNonNull(when);
-      return this;
-    }
-
-    public Builder then(Predicate<Result> then) {
-      this.then = Objects.requireNonNull(then);
-      return this;
-    }
-
-    public Builder assertion(Assertion assertion) {
-      this.assertion = Objects.requireNonNull(assertion);
-      return this;
-    }
-
-    public TestOracle build() {
-      Objects.requireNonNull(when);
-      return new TestOracle() {
-        @Override
-        public Predicate<Tuple> shouldInvoke() {
-          return shouldInvoke;
-        }
-
-        @Override
-        public Function<Tuple, Result> when() {
-          return when;
-        }
-
-        @Override
-        public Predicate<Result> then() {
-          return then;
-        }
-      };
-    }
-  }
-
   interface Result extends Formattable {
     enum Exit {
       RETURNING_VALUE {
