+++ src/test/java/com/github/dakusui/jcunit8/examples/flyingspaghettimonster/FlyingSpaghettiMonsterExample.java
@@ -5,14 +5,14 @@ import com.github.dakusui.jcunit8.factorspace.Parameter.Fsm;
 import com.github.dakusui.jcunit8.factorspace.fsm.Edge;
 import com.github.dakusui.jcunit8.factorspace.fsm.Player;
 import com.github.dakusui.jcunit8.factorspace.fsm.Scenario;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.Condition;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-@RunWith(JCUnit8X.class)
+@RunWith(JCUnit8.class)
 public class FlyingSpaghettiMonsterExample {
   @Condition(constraint = true)
   public boolean check(@From("flyingSpaghettiMonster") Scenario<FlyingSpaghettiMonster> fsm) {
