+++ src/test/java/com/github/dakusui/jcunit8/examples/quadraticequation/QuadraticEquationExample.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.jcunit8.examples.quadraticequation;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter.Simple;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.Condition;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.Given;
@@ -13,7 +13,7 @@ import org.mockito.internal.matchers.LessThan;
 import static java.util.Arrays.asList;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-@RunWith(JCUnit8X.class)
+@RunWith(JCUnit8.class)
 public class QuadraticEquationExample {
   @ParameterSource
   public Simple.Factory<Integer> a() {
