+++ src/test/java/com/github/dakusui/jcunit8/tests/errorhandling/FailOnParameterFactoryCreation.java
@@ -1,13 +1,13 @@
 package com.github.dakusui.jcunit8.tests.errorhandling;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.From;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-@RunWith(JCUnit8X.class)
+@RunWith(JCUnit8.class)
 public class FailOnParameterFactoryCreation {
   public static final String INTENTIONAL_EXCEPTION_MESSAGE = "Intentional exception";
