+++ src/test/java/com/github/dakusui/jcunit8/tests/validation/testclassesundertest/NoTestMethod.java
@@ -1,13 +1,13 @@
 package com.github.dakusui.jcunit8.tests.validation.testclassesundertest;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.runner.RunWith;
 
 import static java.util.Arrays.asList;
 
-@RunWith(JCUnit8X.class)
+@RunWith(JCUnit8.class)
 public class NoTestMethod {
   @ParameterSource
   public Parameter.Simple.Factory<String> simple() {
