+++ src/test/java/com/github/dakusui/jcunit8/tests/validation/testresources/NoParameter.java
@@ -1,10 +1,10 @@
 package com.github.dakusui.jcunit8.tests.validation.testresources;
 
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-@RunWith(JCUnit8X.class)
+@RunWith(JCUnit8.class)
 public class NoParameter {
   @Test
   public void testMethod() {
