+++ src/test/java/com/github/dakusui/jcunit8/tests/validation/testresources/ParameterWithoutFromAnnotation.java
@@ -1,14 +1,14 @@
 package com.github.dakusui.jcunit8.tests.validation.testresources;
 
 import com.github.dakusui.jcunit8.factorspace.Parameter.Simple;
-import com.github.dakusui.jcunit8.runners.junit4.JCUnit8X;
+import com.github.dakusui.jcunit8.runners.junit4.JCUnit8;
 import com.github.dakusui.jcunit8.runners.junit4.annotations.ParameterSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static java.util.Arrays.asList;
 
-@RunWith(JCUnit8X.class)
+@RunWith(JCUnit8.class)
 public class ParameterWithoutFromAnnotation {
   @ParameterSource
   public Simple.Factory<Integer> a() {
